+++ src/main/java/capstone/bookitty/authentication/JwtFilter.java
@@ -1,4 +1,4 @@
-package capstone.bookitty.jwt;
+package capstone.bookitty.authentication;
 
 
 import jakarta.servlet.FilterChain;
