+++ src/main/java/capstone/bookitty/domain/service/MemberService.java
@@ -12,8 +12,8 @@ import capstone.bookitty.domain.entity.RefreshToken;
 import capstone.bookitty.domain.exception.MemberNotFoundException;
 import capstone.bookitty.domain.repository.MemberRepository;
 import capstone.bookitty.domain.repository.RefreshTokenRepository;
-import capstone.bookitty.jwt.JwtToken;
-import capstone.bookitty.jwt.JwtTokenProvider;
+import capstone.bookitty.authentication.JwtToken;
+import capstone.bookitty.authentication.JwtTokenProvider;
 import capstone.bookitty.util.RedisUtil;
 import capstone.bookitty.util.SecurityUtil;
 import jakarta.persistence.EntityNotFoundException;
