+++ src/main/java/capstone/bookitty/error/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package capstone.bookitty.common.error.exception;
+package capstone.bookitty.error.exception;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
